+++ src/main/java/nextstep/subway/line/domain/Line.java
@@ -2,13 +2,9 @@ package nextstep.subway.line.domain;
 
 import nextstep.subway.common.BaseEntity;
 import nextstep.subway.section.domain.Section;
-import nextstep.subway.station.domain.Station;
+import nextstep.subway.section.domain.Sections;
 
 import javax.persistence.*;
-import java.util.ArrayList;
-import java.util.LinkedHashSet;
-import java.util.List;
-import java.util.Set;
 
 @Entity
 public class Line extends BaseEntity {
@@ -22,8 +18,8 @@ public class Line extends BaseEntity {
 
     private String color;
 
-    @OneToMany(mappedBy = "line", cascade = {CascadeType.PERSIST, CascadeType.REMOVE}, orphanRemoval = true)
-    private final List<Section> sections = new ArrayList<>();
+    @Embedded
+    private Sections sections = new Sections();
 
     protected Line() {
     }
@@ -39,10 +35,6 @@ public class Line extends BaseEntity {
         return line;
     }
 
-    public void addSection(Section section) {
-        sections.add(section);
-    }
-
     public Long getId() {
         return id;
     }
@@ -55,12 +47,12 @@ public class Line extends BaseEntity {
         return color;
     }
 
-    public List<Station> getSections() {
-        Set<Station> result = new LinkedHashSet<>();
-        sections.forEach(section -> {
-            result.add(section.getUpStation());
-            result.add(section.getDownStation());
-        });
-        return new ArrayList<>(result);
+    public Sections getSections() {
+        return sections;
     }
+
+    public void addSections(Section section) {
+        this.sections.addSection(section);
+    }
+
 }
