+++ src/main/java/nextstep/subway/line/domain/LineRepository.java
@@ -1,7 +1,8 @@
 package nextstep.subway.line.domain;
 
-import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
@@ -10,12 +11,16 @@ import java.util.Optional;
 @Repository
 public interface LineRepository extends JpaRepository<Line, Long> {
 
-    @Override
-    @EntityGraph(attributePaths = {"sections.upStation","sections.downStation"}, type = EntityGraph.EntityGraphType.LOAD)
-    Optional<Line> findById(Long id);
+    @Query("select l from Line l " +
+            "left join fetch l.sections.sections sec " +
+            "left join fetch sec.downStation " +
+            "left join fetch sec.upStation " +
+            "where l.id = :id")
+    Optional<Line> findOneWithStations(@Param("id") Long id);
 
-    @Override
-    @EntityGraph(attributePaths = {"sections.upStation","sections.downStation"}, type = EntityGraph.EntityGraphType.LOAD)
-    List<Line> findAll();
-
+    @Query("select l from Line l " +
+            "left join fetch l.sections.sections sec " +
+            "left join fetch sec.downStation " +
+            "left join fetch sec.upStation")
+    List<Line> findAllWithStations();
 }
